fixed exit with non numeric arguments + fixed exit with multiple arguments + secured all execution part with const to be sure not to modify something I dont want to change + adding some size_t for the sake of it

+++ exec/utils.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   utils.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fgeorgea <fgeorgea@student.42.fr>          +#+  +:+       +#+        */
+/*   By: fgeorgea <fgeorgea@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/03 12:55:29 by fgeorgea          #+#    #+#             */
-/*   Updated: 2023/05/11 21:28:16 by fgeorgea         ###   ########.fr       */
+/*   Updated: 2023/05/12 01:03:14 by fgeorgea         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
 
-int	arraylen(const char **array)
+size_t	arraylen(const char **array)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (!array || !*array)
